requirements.txt
discord
python-dotenv
audioop-lts
requests
bs4
autokattis
asyncpg
leetcode-api-wrapper